refactor: separate chart config service
